move take_component to Storages and add some safety comments
